requirements.txt
psycopg2-binary>=2.9
clickhouse-connect>=0.7
requests>=2.31
PyYAML>=6.0
python-dotenv>=1.0
